pyproject.toml
[project]
name = "agent-os"
version = "0.1.0"
description = "Personal life OS — markdown-backed, browser UI"
requires-python = ">=3.13"
dependencies = [
    "textual>=0.89.0",
    "pydantic>=2.0.0",
    "python-frontmatter>=1.1.0",
]

[project.scripts]
agent-os = "agent_os.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.15.0"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "B",   # bugbear
    "D300",
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "C90", # mccabe
    "C4",  # comprehensions
    "I",   # isort
    "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
combine-as-imports = true
order-by-type = true
known-first-party = ["agent_os"]
